Update: Add documentation access instructions to README
The README has been updated to include instructions on how to access the API documentation using the Swagger UI interface. Users can now easily explore and interact with the API endpoints by following the provided steps. This addition enhances the user experience and makes it more convenient to understand and use the API.